Add a VRT_HashStrands32 utility
This de-duplicates logic between the two hashing directors, and in
the case of vmod-shard adds a missing check for empty components in
the strands that should be skipped.

This breaks the hashing of the hash director, but unlike the shard
director it offers no stability and results may vary depending on the
health of individual clusters. For this reason only d00003.vtc needed
a bit of reshuffling.
